Split MemModel into logical sections, move MemImpl to its own module
The sections model the organization of the exports, and make
it easier to see what's going on.
